Changed refresh timeout type from in to long in ServerBrowser

+++ JKClient/ServerBrowser.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace JKClient {
 	public sealed class ServerBrowser : NetClient {
-		private const int RefreshTimeout = 3000;
+		private const long RefreshTimeout = 3000L;
 		private readonly List<ServerAddress> masterServers;
 		private readonly ConcurrentDictionary<NetAddress, ServerInfo> globalServers;
 		private TaskCompletionSource<IEnumerable<ServerInfo>> getListTCS, refreshListTCS;
